+++ pymatgen/core/units.py
@@ -187,7 +187,7 @@ def check_mappings(u):
     return u
 
 
-class Unit(collections.Mapping):
+class Unit(collections.abc.Mapping):
     """
     Represents a unit, e.g., "m" for meters, etc. Supports compound units.
     Only integer powers are supported for units.