+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Change log
 ==========
 
+v2019.6.5
+---------
+* Linear scaling get_all_neighbors. Tested to be faster for > 100 atoms (@chc273). 
+* Lobsterin class to handle input for Lobster (@JaGeo).
+* Strict options for composition parsing (@mkhorton).
+* Bug fix for CovalentBondNN.get_bonded_structure (@lan496).
+
 v2019.5.28
 ----------
 * New VASP Input Set "from previous" interface (@utf)